Fix sign-conversion warnings in SampleBroadcaster decompressSamples
Agent-Logs-Url: https://github.com/huberp/phu-beat-sync-multi-scope/sessions/78465baa-889e-4a50-b9a8-d4384aa06d73

+++ lib/network/SampleBroadcaster.cpp
@@ -209,11 +209,11 @@ void SampleBroadcaster::compressSamples(const float* input, int inputBins,
 void SampleBroadcaster::decompressSamples(const uint8_t* input, int numBins,
                                           std::vector<float>& output) {
     const float dbRange = DB_CEILING - DB_FLOOR; // 60 dB
-    output.resize(numBins);
+    output.resize(static_cast<size_t>(numBins));
 
     for (int i = 0; i < numBins; ++i) {
         float normalized = static_cast<float>(input[i]) / 255.0f; // [0, 1]
-        output[i] = normalized * dbRange + DB_FLOOR;               // [-60, 0] dB
+        output[static_cast<size_t>(i)] = normalized * dbRange + DB_FLOOR; // [-60, 0] dB
     }
 }
